Add Ollama advanced engine sharing the batch/context translate loop
The batch/context/schema logic moves from LlamaCppAdvancedTranslate into
AdvancedTranslatorBase; subclasses only supply the endpoint and, where
the server needs one, the model name. The new "Ollama advanced (local
LLM)" engine targets Ollama's OpenAI-compatible /v1/chat/completions
endpoint (stored separately from the classic engine's native
/api/generate URL), reuses the same installed-model list and picker, and
shares the Advanced settings sub-window. Ollama ignores cache_prompt and
recent versions accept response_format json_schema (verified against
gemma3; one cold-start derailment was caught by the existing
retry-then-bisect fallback).
