+++ src/ui/Features/Translate/LlamaCppAdvanced/OllamaAdvancedTranslate.cs
@@ -1,0 +1,29 @@
+using Nikse.SubtitleEdit.Logic.Config;
+
+namespace Nikse.SubtitleEdit.Features.Translate.LlamaCppAdvanced;
+
+/// <summary>
+/// The advanced engine against Ollama's OpenAI-compatible endpoint
+/// (http://localhost:11434/v1/chat/completions - not the native /api/generate the classic
+/// Ollama engine uses). Ollama serves many models, so the "model" field is required. It ignores
+/// llama.cpp's cache_prompt (it manages its own KV cache) and recent versions accept
+/// response_format json_schema; older ones get the json_object/plain fallbacks.
+/// </summary>
+public class OllamaAdvancedTranslate : AdvancedTranslatorBase
+{
+    public static string StaticName { get; set; } = "Ollama advanced (local LLM)";
+    public override string ToString() => StaticName;
+    public override string Name => StaticName;
+    public override string Url => "https://ollama.com";
+
+    protected override string GetApiUrl()
+    {
+        return Se.Settings.AutoTranslate.OllamaAdvancedUrl;
+    }
+
+    protected override string? GetModel()
+    {
+        var model = Se.Settings.AutoTranslate.OllamaAdvancedModel;
+        return string.IsNullOrWhiteSpace(model) ? null : model.Trim();
+    }
+}
