+++ src/ui/Logic/Config/SeAutoTranslate.cs
@@ -33,6 +33,11 @@ public class SeAutoTranslate
     /// </summary>
     public bool LlamaCppUseRemoteServer { get; set; }
     public SeLlamaCppAdvanced LlamaCppAdvanced { get; set; } = new();
+
+    // The Ollama advanced engine uses Ollama's OpenAI-compatible endpoint, not the native
+    // /api/generate URL the classic Ollama engine stores in OllamaUrl.
+    public string OllamaAdvancedUrl { get; set; } = "http://localhost:11434/v1/chat/completions";
+    public string OllamaAdvancedModel { get; set; } = string.Empty;
     public string GroqUrl { get; set; }
     public string GroqPrompt { get; set; }
     public string GroqApiKey { get; set; }
